Filled in missing docstrings.
Minor bugfix where a function that was not used frequently was given a return value.

Major bugfix attempt to messaging-gateway.js related to processing postMessages that are accepted.  In general, it is a tricky issue to mediate between many frames, due to the issue of uniquely identifying a frame (e.g., might have multiple instances of the same child frame URL communicating with a parent).  This must be handled by matching the ID's for the PostMessageStubs and PostMessageGateways.  Added better handling to make this work in more conditions, but still requires testing.