Modified sendCompletedLesson, sendCompletedSublesson, sendCompletedTask, sendCompletedStep to take an optional argument, fake. If fake is true then change verb to terminated.
Modified sendTerminatedSession. If current activity is not session, then call appropriate method to "complete/terminate" activity until all activities are removed from activity tree or session is reached.

These fake completion/termination statements have the context extension: key: https://github.com/GeneralizedLearningUtilities/SuperGLU/recovered, value: activity

This is meant to flag that entries other than object may be problematic.

The session terminated statement has the context extension: https://github.com/GeneralizedLearningUtilities/SuperGLU/missingCompletions, value: True

In the case that there is no session activity in the activity tree, a fake session terminated statement will be sent with the following context extension:

https://github.com/GeneralizedLearningUtilities/SuperGLU/recovered, value: activityPlusContext

This is meant to flag that we are including the context passed to the method as an argument.